Reconstruct modules of `api::vi::accounts::id`
- Remove unnecessary aliasing of `api::v1::statuses::id`
