--- src/syncronous/methods/api/v1/accounts/id/unblock.rs
@@ -1,35 +1,0 @@
-//! This module provides features related to unblock accounts.
-use serde::Serialize;
-use crate::{
-	Connection,
-	Method,
-	entities::Relationship,
-};
-
-/// Get a request to unblock an account specified by `id`.
-pub fn post(conn: &Connection, id: impl Into<String>) -> PostUnblock {
-	PostUnblock {
-		conn,
-		id: id.into(),
-		authorized: true
-	}
-}
-
-/// POST request for `/api/v1/accounts/:id/unblock`.
-#[derive(Debug, Clone, Serialize, mastors_derive::Method)]
-#[method_params(POST, Relationship, "/api/v1/accounts/_PATH_PARAM_/unblock")]
-pub struct PostUnblock<'a> {
-	#[serde(skip_serializing)]
-	#[mastors(connection)]
-	conn: &'a Connection,
-
-	#[serde(skip_serializing)]
-	#[mastors(path_param)]
-	id: String,
-
-	#[serde(skip_serializing)]
-	#[mastors(authorization)]
-	authorized: bool,
-}
-
-impl<'a> Method<'a, Relationship> for PostUnblock<'a> {}
